% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_clean.R
\name{board_clean}
\alias{board_clean}
\title{Remove punctuation from scrabble matrix}
\usage{
board_clean(m)
}
\arguments{
\item{m}{a matrix of scrabble plays}
}
\value{
a matrix of scrabble plays with no notation
}
\description{
Remove punctuation from scrabble matrix
}
\examples{
board_clean(scrab2)
}
