% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_dots.R
\name{board_dots}
\alias{board_dots}
\title{Add dots to scrabble matrix}
\usage{
board_dots(m)
}
\arguments{
\item{m}{a matrix of scrabble plays}
}
\value{
a matrix of scrabble plays with dots where no tiles
}
\description{
Add dots to scrabble matrix
}
\examples{
board_dots(scrab2)
}
