% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_pretty.R
\name{board_pretty}
\alias{board_pretty}
\title{Convert character matrix to pretty scrabble matrix}
\usage{
board_pretty(m)
}
\arguments{
\item{m}{the character matrix}
}
\value{
a square matrix of scrabble plays
}
\description{
Convert character matrix to pretty scrabble matrix
}
\examples{
board_pretty(scrab1)
}
