% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/board_words.R
\name{board_words}
\alias{board_words}
\title{Get all words in a scrabble matrix}
\usage{
board_words(m)
}
\arguments{
\item{m}{a matrix of scrabble plays}
}
\value{
a vector of all words on the scrabble board
}
\description{
Get all words in a scrabble matrix
}
\examples{
board_words(scrab2)
}
