% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards.R
\docType{data}
\name{boards}
\alias{boards}
\title{Boards of scrabble games}
\format{A list of 2566 elements each containing a list  of matrices:}
\usage{
boards
}
\description{
Board position of all scrabble plays from
every turn of each scrabble game
}
\keyword{datasets}
