% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boards_chr.R
\docType{data}
\name{boards_chr}
\alias{boards_chr}
\title{Boards of scrabble games}
\format{A list of 2566 elements each containing a list  of matrices:}
\usage{
boards_chr
}
\description{
Board position of all scrabble plays from
every turn of each scrabble game in character matrix
format
}
\keyword{datasets}
