% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonus_squares.R
\name{bonus_squares}
\alias{bonus_squares}
\title{Determine frequency of open bonus squares}
\usage{
bonus_squares(m)
}
\arguments{
\item{m}{a matrix of scrabble plays with notation}
}
\value{
a dataframe of open bonus squares
}
\description{
Determine frequency of open bonus squares
}
\examples{
bonus_squares(scrab2)
}
