% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalboards.R
\docType{data}
\name{finalboards}
\alias{finalboards}
\title{Final boards of scrabble games}
\format{A list of 2566 matrices:}
\usage{
finalboards
}
\description{
Final board position of all scrabble plays
in each scrabble game
}
\keyword{datasets}
