% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finalboards_chr.R
\docType{data}
\name{finalboards_chr}
\alias{finalboards_chr}
\title{Final boards of scrabble games}
\format{A list of 2566 matrices:}
\usage{
finalboards_chr
}
\description{
Final board position of all scrabble plays
in each scrabble game in character matrix
format
}
\keyword{datasets}
