% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notation.R
\docType{data}
\name{notation}
\alias{notation}
\title{Scrabble board notation}
\format{A data frame with 4 rows and 3 variables:
\describe{
  \item{symbol}{symbol used on board}
  \item{value}{type of bonus square}
  \item{total}{total bonus squares of type on board}
}}
\usage{
notation
}
\description{
Glossary of symbols used to notate scrabble boards
and initial board distribution.
}
\keyword{datasets}
