% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rack_value.R
\name{rack_value}
\alias{rack_value}
\title{Get total value  of a scrabble rack}
\usage{
rack_value(rack)
}
\arguments{
\item{rack}{a string representing scrabble tiles}
}
\value{
integer of total value of all tiles
}
\description{
Get total value  of a scrabble rack
}
\examples{
rack_value("FFJXUUZ")
}
