% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrab1.R
\docType{data}
\name{scrab1}
\alias{scrab1}
\title{Example notated scrabble board}
\format{A 15x15 square matrix:}
\usage{
scrab1
}
\description{
A character matrix of an example scrabble board
including notation for bonus squares.
}
\keyword{datasets}
