% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrab2.R
\docType{data}
\name{scrab2}
\alias{scrab2}
\title{Example notated scrabble board}
\format{A 15x15 square matrix:}
\usage{
scrab2
}
\description{
A matrix of an example scrabble board
including notation for bonus squares.
}
\keyword{datasets}
