% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiles.R
\docType{data}
\name{tiles}
\alias{tiles}
\title{Scrabble tile distribution}
\format{A data frame with 27 rows and 3 variables:
\describe{
  \item{tile}{scrabble tile}
  \item{points}{points value}
  \item{number}{number in game}
}}
\usage{
tiles
}
\description{
Summary of total tiles and their point values
in scrabble game
}
\keyword{datasets}
